% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{esmdata_raw}
\alias{esmdata_raw}
\title{Raw ESM data set}
\format{
A data frame with 6 rows and 13 columns:
\describe{
\item{dyad}{Dyad identification number}
\item{id}{Participant identification number}
\item{role}{Role of a participant within a dyad, indicated by a character (0=father or 1=mother)}
\item{age}{Age of the participant}
\item{scheduled}{Timestamps (e.g., “2023/04/14 10:23:47”) of when the ESM questionnaire was scheduled }
\item{sent}{Timestamps (e.g., “2023/04/14 10:23:47”) of when the ESM questionnaire was sent }
\item{start}{Timestamps (e.g., “2023/04/14 10:23:47”) of when the ESM questionnaire was opened by the participant }
\item{end}{Timestamps (e.g., “2023/04/14 10:23:47”) of when the ESM questionnaire was ended by the participant }
\item{pos_aff}{Positive affect score (0-100)}
\item{neg_aff}{Negative affect score (0-100)}
\item{perc_stress_child}{parents reported if their child did experience a stressful event since the last beep}
\item{perc_fun_child}{parents reported if their child had experienced anything fun since the last beep (yes/no)}
\item{perc_fun_signaled}{parents rated to which extent their child showed they experienced this fun event (0-100)}
}
}
\usage{
esmdata_raw
}
\description{
Raw dataset from a pilot ESM study.
This dataset has undergone random value alterations to ensure privacy and cannot be utilized for formal research purposes.
For additional information, refer to the associated article.
}
\keyword{datasets}
