% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEmstoolsEnv.R
\name{getEsmtoolsEnv}
\alias{getEsmtoolsEnv}
\title{Get the internal package environment}
\usage{
getEsmtoolsEnv()
}
\value{
An environment object used internally by the esmtools package.
}
\description{
This function provides access to the internal environment used by
the esmtools package. This environment is used for internal state
management and should be used with caution.
}
\examples{
esmtools_env <- getEsmtoolsEnv()
print(esmtools_env)
}
