% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dataInfo.R
\name{print.dataInfo}
\alias{print.dataInfo}
\title{Print method of the dataInfo object}
\usage{
\method{print}{dataInfo}(x, ...)
}
\arguments{
\item{x}{A list of data information from the 'dataInfo' function.}

\item{...}{Dots are for compatibility and not used.}
}
\value{
The function returns the kable object invisibly from the 'dataInfo()' function, allowing for its use in further function calls or
command chaining without printing the object again.
}
\description{
This function provides a customized print method for objects created by the 'dataInfo()' function.
It displays each element of the data information list in a user-friendly format.
}
