% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randrows.R
\name{randrows}
\alias{randrows}
\title{Randomly Sample Rows from a Dataframe}
\usage{
randrows(.data, n = 5)
}
\arguments{
\item{.data}{The dataframe from which rows are to be sampled.}

\item{n}{The number of rows to be randomly sampled. Default is 5.}
}
\value{
A dataframe containing the randomly sampled rows.
}
\description{
The \code{randrows()} function randomly samples a specified number of rows from a dataframe.
}
\examples{
# Randomly sample 3 rows from the dataset
randrows(esmdata_sim, n = 3)
}
